perf(ci): shard tests for ≤60s per ubuntu job; PR runs ubuntu-only fast path
Cut the monolithic test job (single ~190s vitest run, gated by a 170s cluster
file; 12-29min on CI) into parallel, wall-time-balanced shards so every job on
the PR critical path stays ≤60s — verified on real CI (48 ubuntu shards run
15-59s).

PR critical path = ubuntu only:
- Test sharding (scripts/run-shard.js, driven by the CI matrix): fork-heavy
  packages split via vitest --shard (cluster→3, egg→2, mock→2, schedule→4,
  security→2, plus redis/multipart/development standalone), each capped with
  --maxWorkers so forked egg cluster processes don't oversubscribe the CPU; the
  light remainder greedy bin-packed by per-dir cpu weight into 8 rest-* shards.
- Long single test files split by `describe` block (no semantic change, no
  removed cases) so vitest can parallelize them.
- quality split into a parallel matrix (lint/typecheck/fmtcheck/build/site) so
  the slowest check — not their sum — bounds wall time.
- egg-bin runs without --coverage to stay ≤60s (no threshold; report only).

macOS/Windows whole-suite compatibility (slow: ~13-25min; macOS has only ~5
runner slots) moved to a `test-compat` job that runs at merge time
(merge_group) and on push to next — NOT on pull_request — so it no longer gates
the PR wall while still enforcing cross-platform coverage before code lands.
CI test jobs on a PR drop from 132 to the 48 ubuntu shards.

Latent bugs fixed along the way:
- tegg controller ControllerMetaManager.test.ts triggers a duplicate-proto
  error that polluted the process-global registry under isolate:false, failing
  sibling files; isolated that project (isolate:true).
- clean-dist used `ut run clean --workspaces` which aborts the && chain (root has
  no clean script) → scripts/clean-dist.js.
- typecheck/pretest used `ut run <self> --workspaces`, which includes the root
  and recurses infinitely → scripts/run-workspaces.js (concurrent, root-bin PATH).
- Windows runner images no longer set %HOME%, breaking egg_loader getHomedir;
  export HOME=%USERPROFILE% on the windows compat job.

Windows flakes absorbed by --retry. No test semantics changed or cases removed.
