+++ package.json
@@ -13,19 +13,21 @@
   ],
   "type": "module",
   "scripts": {
-    "clean-dist": "ut run clean --workspaces --if-present",
+    "clean-dist": "node scripts/clean-dist.js",
     "build": "tsdown",
     "prelint": "ut run clean-dist",
     "lint": "oxlint --type-aware --type-check --quiet",
     "fmt": "oxfmt",
-    "typecheck": "ut run clean-dist && ut run typecheck --workspaces --if-present",
+    "typecheck": "ut run clean-dist && node scripts/run-workspaces.js typecheck",
     "fmtcheck": "oxfmt --check .",
-    "pretest": "ut run clean-dist && ut run pretest --workspaces --if-present",
+    "pretest": "ut run clean-dist && node scripts/run-workspaces.js pretest",
     "test": "vitest run --bail 1 --retry 2 --testTimeout 20000 --hookTimeout 20000",
     "test:cov": "ut run test -- --coverage",
+    "test:shard": "node scripts/run-shard.js",
     "benchmark:ci-test": "node scripts/ci-test-benchmark.js",
-    "preci": "ut run pretest --workspaces --if-present",
+    "preci": "ut run pretest",
     "ci": "ut run test -- --coverage",
+    "ci:shard": "node scripts/run-shard.js --coverage",
     "dev:services:start": "node scripts/dev-services.js start",
     "dev:services:stop": "node scripts/dev-services.js stop",
     "dev:services:status": "node scripts/dev-services.js status",
