+++ packages/cluster/test/https.test.ts
@@ -67,26 +67,4 @@ describe('test/https.test.ts', () => {
       assert(response.data === 'https server');
     });
   });
-
-  describe('start https server with app config cluster', () => {
-    afterEach(() => app && app.close());
-
-    it('should success with status 200', async () => {
-      const port = await detectPort();
-      const options = {
-        baseDir: getFilepath('apps/https-server-config'),
-        port,
-      };
-
-      app = cluster('apps/https-server-config', options);
-      await app.ready();
-
-      const response = await httpclient.request(`https://127.0.0.1:${port}`, {
-        dataType: 'text',
-      });
-
-      assert(response.status === 200);
-      assert(response.data === 'https server config');
-    });
-  });
 });
