+++ packages/egg/test/lib/plugins/watcher.test.ts
@@ -3,7 +3,7 @@ import fs from 'node:fs';
 import { scheduler } from 'node:timers/promises';
 
 import { mm } from '@eggjs/mock';
-import { describe, it, beforeEach, afterEach, beforeAll, afterAll } from 'vitest';
+import { describe, it, beforeEach, afterEach } from 'vitest';
 
 import { cluster, type MockApplication, getFilepath } from '../../utils.ts';
 
@@ -72,24 +72,6 @@ describe('test/lib/plugins/watcher.test.ts', () => {
           count = parseInt(res.text);
           assert(count > lastCount);
         });
-    });
-  });
-
-  describe('config.watcher.type is default', () => {
-    let app: MockApplication;
-    beforeAll(() => {
-      app = cluster('apps/watcher-type-default');
-      app.coverage(false);
-      return app.ready();
-    }, 60000);
-
-    afterAll(() => app.close());
-
-    it('should warn user', async () => {
-      await scheduler.wait(3000);
-      const logPath = getFilepath('apps/watcher-type-default/logs/watcher-type-default/egg-agent.log');
-      const content = fs.readFileSync(logPath, 'utf8');
-      assert.match(content, /defaultEventSource watcher will NOT take effect/);
     });
   });
 });
