+++ plugins/schedule/test/immediate.test.ts
@@ -24,25 +24,3 @@ describe.skipIf(process.platform === 'win32')('cluster - immediate', () => {
     expect(contains(log, 'immediate-cron')).toBeGreaterThanOrEqual(2);
   });
 });
-
-describe.skipIf(process.platform === 'win32')('cluster - immediate-onlyonce', () => {
-  let app: MockApplication;
-  beforeAll(async () => {
-    app = mm.cluster({
-      baseDir: getFixtures('immediate-onlyonce'),
-      workers: 1,
-    });
-    // app.debug();
-    await app.ready();
-  });
-  afterAll(() => app.close());
-
-  it('should work', async () => {
-    await sleep(5000);
-
-    const log = getLogContent('immediate-onlyonce');
-    // console.log(log);
-    // unstable
-    expect(contains(log, 'immediate-onlyonce')).toBeGreaterThanOrEqual(0);
-  });
-});
