+++ plugins/schedule/test/schedule-type-worker-ctxstorage.test.ts
@@ -1,0 +1,33 @@
+import { setTimeout as sleep } from 'node:timers/promises';
+
+import { mm, type MockApplication } from '@eggjs/mock';
+import { describe, it, afterAll, beforeAll, expect } from 'vitest';
+
+import { contains, getFixtures, getLogContent, getScheduleLogContent } from './utils.ts';
+
+describe('cluster - worker-ctxStorage', () => {
+  let app: MockApplication;
+  beforeAll(async () => {
+    app = mm.cluster({ baseDir: getFixtures('worker-ctxStorage'), workers: 1 });
+    // app.debug();
+    await app.ready();
+  });
+  afterAll(() => app.close());
+
+  it('should work', async () => {
+    await sleep(5000);
+    const log = getLogContent('worker-ctxStorage');
+    // console.log(log);
+    // unstable
+    expect(contains(log, 'interval')).toBeGreaterThanOrEqual(0);
+    expect(contains(log, 'foobar')).toBeGreaterThanOrEqual(0);
+
+    const scheduleLog = getScheduleLogContent('worker-ctxStorage');
+    // console.log(scheduleLog);
+    // unstable
+    expect(contains(scheduleLog, 'foobar.js executing by app')).toBeGreaterThanOrEqual(0);
+    expect(contains(scheduleLog, 'foobar.js execute succeed')).toBeGreaterThanOrEqual(0);
+    expect(contains(scheduleLog, 'interval.js executing by app')).toBeGreaterThanOrEqual(0);
+    expect(contains(scheduleLog, 'interval.js execute succeed')).toBeGreaterThanOrEqual(0);
+  });
+});
